Fix some integer size warnings in hsStream

+++ Sources/Plasma/FeatureLib/pfPatcher/pfPatcher.cpp
@@ -280,7 +280,7 @@ public:
         fOutput.reset();
     }
 
-    uint32_t Write(uint32_t count, const void* buf) override
+    size_t Write(size_t count, const void* buf) override
     {
         // tick whatever progress bar we have
         IUpdateProgress(count);
@@ -295,7 +295,7 @@ public:
     bool AtEnd() override { return fOutput->AtEnd(); }
     uint32_t GetEOF() override { return fOutput->GetEOF(); }
     uint32_t GetPosition() const override { return fOutput->GetPosition(); }
-    uint32_t Read(uint32_t count, void* buf) override { return fOutput->Read(count, buf); }
+    size_t Read(size_t count, void* buf) override { return fOutput->Read(count, buf); }
     void Rewind() override { fOutput->Rewind(); }
     void FastFwd() override { fOutput->FastFwd(); }
     void SetPosition(uint32_t pos) override { fOutput->SetPosition(pos); }
